♻️ refactor: rename the file names to remove redundant `_models` suffix (#623)
