+++ src/models/mod.rs
@@ -2,7 +2,7 @@
 //! results, parsing config file, providing trait to standardize search engine handling code,
 //! custom engine error for the search engine, etc.
 
-pub mod aggregation_models;
-pub mod engine_models;
-pub mod parser_models;
-pub mod server_models;
+pub mod aggregation;
+pub mod engine;
+pub mod parser;
+pub mod search_route;
